Include exoplayer and bintray scripts (#3)
* add exoplayer libraries

* include exoplayer module

* include gradle scripts

* cleaning code and fixing behaviors

+++ domain/src/main/kotlin/br/com/suamusica/rxmediaplayer/domain/MediaProgress.kt
@@ -1,8 +1,8 @@
 package br.com.suamusica.rxmediaplayer.domain
 
-data class MediaProgress(val current: Int, val total: Int) {
+data class MediaProgress(val current: Long, val total: Long) {
 
-  fun percentage(): Int = (current * 100) / total
+  fun percentage(): Long = (current * 100) / total
 
   companion object {
     val COMPLETED = MediaProgress(100, 100)
